fix invariant chemical potentials
assign names to chemical potentials and chemical potential parameters, otherwise there is a segfault when an error occurs

+++ mpi_utils.c
@@ -110,7 +110,7 @@ void broadcast_params()
   blklens[4] = SLOTS;     typen[4] = MPI_INT;     /* slot */
   blklens[5] = SLOTS;     typen[5] = REAL;        /* shift */
   blklens[6] = SLOTS;     typen[6] = REAL;        /* step */
-  blklens[7] = SLOTS;     typen[7] = REAL;        /* col */
+  blklens[7] = SLOTS;     typen[7] = MPI_INT;     /* col */
   size = 8;
 #ifdef ADP
   blklens[8] = 1;         typen[8] = MPI_VEKTOR;  /* rdist */
